moves separable table component to own folder
