+++ src/admin/components/table/table-entry.js
@@ -1,0 +1,9 @@
+var React = require('react')
+
+var TableEntry = module.exports = React.createClass({
+  render: function () {
+    return (
+      <div className='table-entry'>
+        <p className={ this.props.header } > { this.props.entry } </p>
+      </div> )}})
+
